Update: 반려견 정보 수정 기능 - 반려견 이미지 수정시 로컬에 저장되어 있던 기존 이미지 파일을 삭제하는 기능 추가

+++ Goodee-Semi/src/main/java/com/goodee/semi/controller/MyPetUpdateServlet.java
@@ -111,19 +111,23 @@ public class MyPetUpdateServlet extends HttpServlet {
     				attach.setTypeNo(Attach.PET);
     				attach.setPkNo(pet.getPetNo());
     				
+    				// TODO: 아래의 기능 마저 구현
     				// 기존 이미지파일 삭제
     				// 1. petNo로 기존 이미지파일의 savedName을 가져옴
     				// 2. savedName에 해당하는 파일을 저장소에서 삭제
-    				String oldSavedName = petService.selectPetImgSavedName(pet.getPetNo());
+    				String oldSavedName = petService.selectPetImgSavedName(attach);
     				System.out.println("[MyPetUpdateServlet] 삭제할 파일명: " + oldSavedName);
-    				File oldSavedFile = new File(AttachService.getUploadDirectory(Attach.PET) + "/" + oldSavedName);
+    				File oldSavedFile = new File(AttachService.getUploadDirectory(Attach.PET), oldSavedName);
     				System.out.println("[MyPetUpdateServlet] 삭제할 파일경로: " + oldSavedFile.getAbsolutePath());
-    				if (oldSavedFile.delete()) {
-    					System.out.println("[MyPetUpdateServlet] 기존 반려견 이미지 삭제 성공");
+    				if(oldSavedFile.exists()) {
+    					if (oldSavedFile.delete()) {
+    						System.out.println("[MyPetUpdateServlet] 기존 반려견 이미지 삭제에 성공했습니다");
+    					} else {
+    						System.out.println("[MyPetUpdateServlet] 기존 반려견 이미지 삭제에 실패했습니다");
+    					}
     				} else {
-    					System.out.println("[MyPetUpdateServlet] 기존 반려견 이미지 삭제 실패");
+    					System.out.println("[MyPetUpdateServlet] 기존 반려견 이미지 파일이 존재하지 않습니다");
     				}
-    				
     			} catch(Exception e) {
     	            sendErrorResponse("500", "파일 저장/삭제 중 문제가 발생했습니다: " + e.getMessage(), e, response);
     	            return;
